refactor: rename import paths from "./filters/[...] to ./filters
replace to work with index.ts file
