+++ server/src/controllers/messagesController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { RequestQueryMessage } from "@types";
-import { filterMessagesByUrlParams } from "./filters/messagesFilter";
+import { filterMessagesByUrlParams } from "./filters";
 import { getMessages, getMessagesCount } from "@services";
 
 export const getMessagesList = async (
